moved user_self_service_role as global fixture
